pyproject.toml
[tool.poetry]
name = "tatl"
version = "0.1.0"
description = ""
authors = ["Aria Taylor <ari@aricodes.net>"]
readme = "README.md"

[tool.poetry.scripts]
watch = "tatl.watch:main"
tatl = "tatl:main"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
twitchio = "^2.10.0"
py-cord = "^2.6.1"
msgspec = "^0.19.0"
peewee = "^3.17.8"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
watchfiles = "^1.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
